CCD roles junit refactoring
