+++ src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/drools/StaffCategoryCaseTest.java
@@ -8,7 +8,6 @@ import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.roleassignment.domain.model.ExistingRoleAssignment;
 import uk.gov.hmcts.reform.roleassignment.domain.model.FeatureFlag;
 import uk.gov.hmcts.reform.roleassignment.domain.model.RoleAssignment;
-import uk.gov.hmcts.reform.roleassignment.domain.model.enums.Classification;
 import uk.gov.hmcts.reform.roleassignment.domain.model.enums.FeatureFlagEnum;
 import uk.gov.hmcts.reform.roleassignment.domain.model.enums.RoleCategory;
 import uk.gov.hmcts.reform.roleassignment.domain.model.enums.Status;
@@ -17,7 +16,6 @@ import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static uk.gov.hmcts.reform.roleassignment.domain.model.enums.GrantType.CHALLENGED;
@@ -494,167 +492,6 @@ class StaffCategoryCaseTest extends DroolBase {
         assignmentRequest.getRequestedRoles().forEach(roleAssignment -> {
             assertEquals(Status.DELETE_APPROVED, roleAssignment.getStatus());
         });
-    }
-
-    @Test
-    void shouldApproveCaseRequestedRoles_CCD_1_0_NotFromCCD() {
-        RoleAssignment requestedRole1 = getRequestedCaseRole(RoleCategory.PROFESSIONAL, "[PETSOLICITOR]",
-                                                             SPECIFIC, "caseId",
-                                                             "1234567890123456", CREATE_REQUESTED);
-        assignmentRequest.setRequestedRoles(List.of(requestedRole1));
-        FeatureFlag featureFlag  =  FeatureFlag.builder().flagName(FeatureFlagEnum.CCD_1_0.getValue())
-            .status(true).build();
-        featureFlags.add(featureFlag);
-
-        executeDroolRules(List.of(buildExistingRoleForIAC(requestedRole1.getActorId(),
-                                                          "senior-tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "[PETSOLICITOR]")));
-        //assertion
-        assignmentRequest.getRequestedRoles().forEach(ra -> assertEquals(Status.REJECTED, ra.getStatus()));
-    }
-
-    @Test
-    void shouldApproveCreaterCaseRequestedRoles_CCD_1_0_FromCCD() {
-        verifyCreateCaseRequestedRole_CCD_1_0("[CREATOR]", "ccd_data", RoleCategory.JUDICIAL);
-        verifyCreateCaseRequestedRole_CCD_1_0("[CREATOR]", "aac_manage_case_assignment",
-                                              RoleCategory.LEGAL_OPERATIONS);
-    }
-
-    @Test
-    void shouldApproveCreateCaseRequestedRoles_CCD_1_0_FromCCD() {
-        verifyCreateCaseRequestedRole_CCD_1_0("[PETSOLICITOR]", "ccd_data", RoleCategory.PROFESSIONAL);
-        verifyCreateCaseRequestedRole_CCD_1_0("[CREATOR]", "aac_manage_case_assignment",
-                                              RoleCategory.CITIZEN);
-    }
-
-    private void verifyCreateCaseRequestedRole_CCD_1_0(String roleName, String clientId, RoleCategory category) {
-        RoleAssignment requestedRole1 = getRequestedCaseRole(category, roleName,
-                                                             SPECIFIC, "caseId",
-                                                             "1234567890123456", CREATE_REQUESTED);
-        requestedRole1.setClassification(Classification.RESTRICTED);
-        requestedRole1.getAttributes().putAll(Map.of("jurisdiction", convertValueJsonNode("IA"),
-                                                     "caseType", convertValueJsonNode("Asylum"),
-                                                     "caseId", convertValueJsonNode("1234567890123456")));
-        assignmentRequest.setRequestedRoles(List.of(requestedRole1));
-        assignmentRequest.getRequest().setClientId(clientId);
-
-        FeatureFlag featureFlag  =  FeatureFlag.builder().flagName(FeatureFlagEnum.CCD_1_0.getValue())
-            .status(true).build();
-        featureFlags.add(featureFlag);
-
-        executeDroolRules(List.of(buildExistingRoleForIAC(requestedRole1.getActorId(),
-                                                          "senior-tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(), roleName)));
-        //assertion
-        assignmentRequest.getRequestedRoles().forEach(ra -> assertEquals(Status.APPROVED, ra.getStatus()));
-    }
-
-    @Test
-    void shouldApproveDefaultCreateCaseRequestedRoles_CCD_1_0_FromCCD() {
-        verifyDefaultCaseRequestRoleCreate_CCD_1_0(true, Status.APPROVED);
-        verifyDefaultCaseRequestRoleCreate_CCD_1_0(false, Status.REJECTED);
-    }
-
-    private void verifyDefaultCaseRequestRoleCreate_CCD_1_0(Boolean enableByPass, Status status) {
-        RoleAssignment requestedRole1 = getRequestedCaseRole(RoleCategory.SYSTEM, "[RESPSOLICITOR]",
-                                                             SPECIFIC, "caseId",
-                                                             "1234567890123456", CREATE_REQUESTED);
-        requestedRole1.setClassification(Classification.RESTRICTED);
-        requestedRole1.getAttributes().putAll(Map.of("jurisdiction", convertValueJsonNode("IA"),
-                                                     "caseType", convertValueJsonNode("Asylum"),
-                                                     "caseId", convertValueJsonNode("1234567890123456")));
-        assignmentRequest.setRequestedRoles(List.of(requestedRole1));
-        assignmentRequest.getRequest().setClientId("ccd_data");
-        assignmentRequest.getRequest().setByPassOrgDroolRule(enableByPass);
-
-        FeatureFlag featureFlag  =  FeatureFlag.builder().flagName(FeatureFlagEnum.CCD_1_0.getValue())
-            .status(true).build();
-        featureFlags.add(featureFlag);
-
-        executeDroolRules(List.of(buildExistingRoleForIAC(requestedRole1.getActorId(),
-                                                          "senior-tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "[RESPSOLICITOR]")));
-        //assertion
-        assignmentRequest.getRequestedRoles().forEach(ra -> assertEquals(status, ra.getStatus()));
-    }
-
-    @Test
-    void shouldApproveDeleteCaseRequestedRoles_CCD_1_0_FromCCD() {
-        verifyDeleteCaseRequestRole_CCD_1_0("[PETSOLICITOR]", "ccd_data", RoleCategory.PROFESSIONAL);
-        verifyDeleteCaseRequestRole_CCD_1_0("[CREATOR]", "aac_manage_case_assignment",
-                                            RoleCategory.CITIZEN);
-    }
-
-    @Test
-    void shouldApproveDeleteCreaterCaseRequestedRoles_CCD_1_0_FromCCD() {
-        verifyDeleteCaseRequestRole_CCD_1_0("[CREATOR]", "ccd_data", RoleCategory.JUDICIAL);
-        verifyDeleteCaseRequestRole_CCD_1_0("[CREATOR]", "aac_manage_case_assignment",
-                                            RoleCategory.LEGAL_OPERATIONS);
-    }
-
-    private void verifyDeleteCaseRequestRole_CCD_1_0(String roleName, String clientId, RoleCategory category) {
-        RoleAssignment requestedRole1 = getRequestedCaseRole(category, roleName,
-                                                             SPECIFIC, "caseId",
-                                                             "1234567890123456", DELETE_REQUESTED);
-        requestedRole1.setClassification(Classification.RESTRICTED);
-        requestedRole1.getAttributes().putAll(Map.of("jurisdiction", convertValueJsonNode("IA"),
-                                                     "caseType", convertValueJsonNode("Asylum"),
-                                                     "caseId", convertValueJsonNode("1234567890123456")));
-        assignmentRequest.setRequestedRoles(List.of(requestedRole1));
-        assignmentRequest.getRequest().setClientId(clientId);
-
-        FeatureFlag featureFlag  =  FeatureFlag.builder().flagName(FeatureFlagEnum.CCD_1_0.getValue())
-            .status(true).build();
-        featureFlags.add(featureFlag);
-
-        executeDroolRules(List.of(buildExistingRoleForIAC(requestedRole1.getActorId(),
-                                                          "senior-tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          roleName)));
-        //assertion
-        assignmentRequest.getRequestedRoles().forEach(ra -> assertEquals(Status.DELETE_APPROVED, ra.getStatus()));
-    }
-
-    @Test
-    void shouldApproveDefaultDeleteCaseRequestedRoles_CCD_1_0_FromCCD() {
-        verifyDefaultCaseRequestRoleDelete_CCD_1_0(true, Status.DELETE_APPROVED);
-        verifyDefaultCaseRequestRoleDelete_CCD_1_0(false, Status.DELETE_REJECTED);
-    }
-
-    private void verifyDefaultCaseRequestRoleDelete_CCD_1_0(Boolean enableByPass, Status status) {
-        RoleAssignment requestedRole1 = getRequestedCaseRole(RoleCategory.SYSTEM, "[RESPSOLICITOR]",
-                                                             SPECIFIC, "caseId",
-                                                             "1234567890123456", DELETE_REQUESTED);
-        requestedRole1.setClassification(Classification.RESTRICTED);
-        requestedRole1.getAttributes().putAll(Map.of("jurisdiction", convertValueJsonNode("IA"),
-                                                     "caseType", convertValueJsonNode("Asylum"),
-                                                     "caseId", convertValueJsonNode("1234567890123456")));
-        assignmentRequest.setRequestedRoles(List.of(requestedRole1));
-        assignmentRequest.getRequest().setClientId("ccd_data");
-        assignmentRequest.getRequest().setByPassOrgDroolRule(enableByPass);
-
-        FeatureFlag featureFlag  =  FeatureFlag.builder().flagName(FeatureFlagEnum.CCD_1_0.getValue())
-            .status(true).build();
-        featureFlags.add(featureFlag);
-
-        executeDroolRules(List.of(buildExistingRoleForIAC(requestedRole1.getActorId(),
-                                                          "senior-tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "tribunal-caseworker"),
-                                  buildExistingRoleForIAC(assignmentRequest.getRequest().getAssignerId(),
-                                                          "[RESPSOLICITOR]")));
-        //assertion
-        assignmentRequest.getRequestedRoles().forEach(ra -> assertEquals(status, ra.getStatus()));
     }
 
     private void executeDroolRules(List<ExistingRoleAssignment> existingRoleAssignments) {
